Fix CSRF issue
Create a session cookie containing a CSRF token each time a user accesses the home page.
The CSRF token is passed as a query parameter in any request to /goapp/ws.
The CSRF token in the session cookie is compared with the CSRF token retrieved from the query parameter of the request.
The connection is allowed only if these two values are identical.